requirements.txt
discord.py
discord 
Flask 
bs4 
requests
firebase-admin
emoji
aiohttp
async-timeout
Pillow 
asyncio
enka.py
pytz
pandas
essential_generators
matplotlib
psutil
google-genai
beautifulsoup4
matplotlib
numpy
seaborn
